+++ inc/classes/database-classes.php
@@ -180,6 +180,9 @@ class dbHelper {
       ],
       '0.7.7' => [
         "DELETE FROM pages WHERE Url = 'core/pages';"
+      ],
+      '0.7.8' => [
+        "ALTER TABLE pages ADD COLUMN isDefault BOOLEAN;"
       ]
     ];
   }